provider: carve openai.rs (pure move)
The OpenAI Responses-API region moves verbatim out of provider/mod.rs:
request/response types, request-part assembly, the ChatProvider impl,
and CU action parsing, plus the 18 tests that exercise them. Moved
items widen to pub(crate); ReasoningConfig's fields widen too
(resolve_reasoning in mod.rs constructs it across the new boundary).
tool_msg_with_images stays in mod.rs tests as a shared fixture.
